Backend/requirements.txt
Django~=6.0.2
python-dotenv
djangorestframework
djangorestframework-stubs
torch==2.10.0+cu130
lightning
torchmetrics
django-cors-headers